Fix errors

+++ ray-operator/test/e2eautoscaler/raycluster_autoscaler_test.go
@@ -383,7 +383,7 @@ func TestRayClusterAutoscalerMaxReplicasUpdate(t *testing.T) {
 				WithWorkerGroupSpecs(rayv1ac.WorkerGroupSpec().
 					WithReplicas(0).
 					WithMinReplicas(0).
-					WithMaxReplicas(maxReplicas).
+					WithMaxReplicas(int32(maxReplicas)).
 					WithGroupName(groupName).
 					WithRayStartParams(map[string]string{"num-cpus": "1"}).
 					WithTemplate(tc.WorkerPodTemplateGetter()))
@@ -418,7 +418,7 @@ func TestRayClusterAutoscalerMaxReplicasUpdate(t *testing.T) {
 			// Update maxReplicas to half the previous value
 			rayCluster, err = test.Client().Ray().RayV1().RayClusters(namespace.Name).Get(test.Ctx(), rayCluster.Name, metav1.GetOptions{})
 			g.Expect(err).NotTo(gomega.HaveOccurred())
-			rayCluster.Spec.WorkerGroupSpecs[0].MaxReplicas = ptr.To(maxReplicas / 2)
+			rayCluster.Spec.WorkerGroupSpecs[0].MaxReplicas = ptr.To(int32(maxReplicas / 2))
 			rayCluster, err = test.Client().Ray().RayV1().RayClusters(namespace.Name).Update(test.Ctx(), rayCluster, metav1.UpdateOptions{})
 			g.Expect(err).NotTo(gomega.HaveOccurred())
 			test.T().Logf("Updated RayCluster %s/%s successfully", rayCluster.Namespace, rayCluster.Name)
